requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.23.2
sqlalchemy==2.0.43
psycopg2-binary==2.9.7
pydantic==2.11.7
python-dotenv==1.1.1
passlib[bcrypt]==1.7.4
python-jose==3.3.0
email-validator==2.2.0
anyio==4.10.0
starlette==0.47.2
dnspython==2.7.0
